// apps/designer/src/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      instances: {
        Row: {
          account_id: string | null
          company_summary: string | null
          config: Json | null
          created_at: string | null
          credit_price: number
          demo_instance: boolean
          demo_instance_type:
            | Database["public"]["Enums"]["demo_instance_type_enum"]
            | null
          description: string | null
          email_lead_price: number
          flow_config: Json | null
          id: string
          instance_type: string | null
          is_public: boolean | null
          max_submissions_per_session: number
          name: string
          phone_lead_price: number
          slug: string
          submission_limit_enabled: boolean
          updated_at: string | null
          use_case: Database["public"]["Enums"]["use_case_type"]
          user_id: string | null
          webhook_url: string | null
          website_url: string | null
        }
        Insert: {
          account_id?: string | null
          company_summary?: string | null
          config?: Json | null
          created_at?: string | null
          credit_price?: number
          demo_instance?: boolean
          demo_instance_type?:
            | Database["public"]["Enums"]["demo_instance_type_enum"]
            | null
          description?: string | null
          email_lead_price?: number
          flow_config?: Json | null
          id?: string
          instance_type?: string | null
          is_public?: boolean | null
          max_submissions_per_session?: number
          name: string
          phone_lead_price?: number
          slug: string
          submission_limit_enabled?: boolean
          updated_at?: string | null
          use_case: Database["public"]["Enums"]["use_case_type"]
          user_id?: string | null
          webhook_url?: string | null
          website_url?: string | null
        }
        Update: {
          account_id?: string | null
          company_summary?: string | null
          config?: Json | null
          created_at?: string | null
          credit_price?: number
          demo_instance?: boolean
          demo_instance_type?:
            | Database["public"]["Enums"]["demo_instance_type_enum"]
            | null
          description?: string | null
          email_lead_price?: number
          flow_config?: Json | null
          id?: string
          instance_type?: string | null
          is_public?: boolean | null
          max_submissions_per_session?: number
          name?: string
          phone_lead_price?: number
          slug?: string
          submission_limit_enabled?: boolean
          updated_at?: string | null
          use_case?: Database["public"]["Enums"]["use_case_type"]
          user_id?: string | null
          webhook_url?: string | null
          website_url?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "instances_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never